Remove index-vi.html, pending restructuring for multi-language support
This commit removes the index-vi.html file as it is no longer needed. Currently, there is no dedicated support for translating the website into Vietnamese in the latest version. We plan to restructure and implement multi-language support in the near future. Removing the index-vi.html file is part of the cleanup process to streamline the project and prepare for future language enhancements.


Signed-off-by: bTeam Developer <57291312+bteamapp@users.noreply.github.com>