Create a simple 404 error page in English. Update 404.html
Create a simple 404 error page in English

Added a basic HTML file (404.html) for a 404 error page with a large "404" heading and a friendly message. This page serves as a user-friendly response when a requested page is not found. Users can customize the styles and content to fit their preferences.


Signed-off-by: bTeam Developer <57291312+bteamapp@users.noreply.github.com>