Add a simple 500 error page
Add a simple 500 error page

Created a basic HTML page (500.html) to serve as the custom error page for 500 Internal Server Errors. Configured the .htaccess file to point to this error page using the ErrorDocument directive. The error page provides a user-friendly message indicating that something went wrong on the server and that the issue is being addressed. This commit improves the user experience in the event of a 500 error.


Signed-off-by: bTeam Developer <57291312+bteamapp@users.noreply.github.com>